+++ packages/plugin-react/src/index.ts
@@ -119,7 +119,7 @@ export default function viteReact(opts: Options = {}): Plugin[] {
   let projectRoot = process.cwd()
   let skipFastRefresh = true
   let base: string
-  let isFullBundle = false
+  let isBundledDev = false
   let runPluginOverrides:
     | ((options: ReactBabelOptions, context: ReactBabelHookContext) => void)
     | undefined
@@ -163,7 +163,7 @@ export default function viteReact(opts: Options = {}): Plugin[] {
               jsxRefreshExclude: makeIdFiltersToMatchWithQuery(exclude),
             },
             optimizeDeps: {
-              rollupOptions: { transform: { jsx: { runtime: 'automatic' } } },
+              rolldownOptions: { transform: { jsx: { runtime: 'automatic' } } },
             },
           }
         }
@@ -191,8 +191,8 @@ export default function viteReact(opts: Options = {}): Plugin[] {
       base = config.base
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment -- use ts-ignore for ecosystem-ci
       // @ts-ignore only available in newer rolldown-vite
-      if (config.experimental.fullBundleMode) {
-        isFullBundle = true
+      if (config.experimental.bundledDev) {
+        isBundledDev = true
       }
       projectRoot = config.root
       isProduction = config.isProduction
@@ -450,12 +450,12 @@ export default function viteReact(opts: Options = {}): Plugin[] {
   }
 
   // for rolldown-vite + full bundle mode
-  const viteReactRefreshFullBundleMode: Plugin = {
+  const viteReactRefreshBundledDevMode: Plugin = {
     name: 'vite:react-refresh-fbm',
     enforce: 'pre',
     transformIndexHtml: {
       handler() {
-        if (!skipFastRefresh && isFullBundle)
+        if (!skipFastRefresh && isBundledDev)
           return [
             {
               tag: 'script',
@@ -514,7 +514,7 @@ export default function viteReact(opts: Options = {}): Plugin[] {
       },
     },
     transformIndexHtml() {
-      if (!skipFastRefresh && !isFullBundle)
+      if (!skipFastRefresh && !isBundledDev)
         return [
           {
             tag: 'script',
@@ -528,12 +528,12 @@ export default function viteReact(opts: Options = {}): Plugin[] {
   return [
     viteBabel,
     ...(isRolldownVite
-      ? [viteRefreshWrapper, viteConfigPost, viteReactRefreshFullBundleMode]
+      ? [viteRefreshWrapper, viteConfigPost, viteReactRefreshBundledDevMode]
       : []),
     viteReactRefresh,
     virtualPreamblePlugin({
       name: '@vitejs/plugin-react/preamble',
-      isEnabled: () => !skipFastRefresh && !isFullBundle,
+      isEnabled: () => !skipFastRefresh && !isBundledDev,
     }),
   ]
 }