+++ packages/plugin-rsc/examples/no-ssr/src/framework/entry.browser.tsx
@@ -57,7 +57,9 @@ async function main() {
       { temporaryReferences },
     )
     setPayload(payload)
-    return payload.returnValue
+    const { ok, data } = payload.returnValue!
+    if (!ok) throw data
+    return data
   })
 
   // hydration
@@ -68,7 +70,7 @@ async function main() {
   )
   createRoot(document.body).render(browserRoot)
 
-  // implement server HMR by trigering re-fetch/render of RSC upon server code change
+  // implement server HMR by triggering re-fetch/render of RSC upon server code change
   if (import.meta.hot) {
     import.meta.hot.on('rsc:update', () => {
       fetchRscPayload()