+++ packages/plugin-react/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+## 5.1.2 (2025-12-08)
+
+## 5.1.1 (2025-11-12)
+
+### Update code to support newer `rolldown-vite` ([#976](https://github.com/vitejs/vite-plugin-react/pull/976))
+
+`rolldown-vite` will remove `optimizeDeps.rollupOptions` in favor of `optimizeDeps.rolldownOptions` soon. This plugin now uses `optimizeDeps.rolldownOptions` to support newer `rolldown-vite`. Please update `rolldown-vite` to the latest version if you are using an older version.
+
 ## 5.1.0 (2025-10-24)
 
 ### Add `@vitejs/plugin-react/preamble` virtual module for SSR HMR ([#890](https://github.com/vitejs/vite-plugin-react/pull/890))